+++ A32NX/html_ui/Pages/VCockpit/Instruments/Airliners/A320_Neo/FMC/A32NX_FMCMainDisplay.js
@@ -1277,7 +1277,19 @@ class FMCMainDisplay extends BaseAirliners {
     }
 
     checkVSpeedDisagree(mcdu) {
-        return mcdu.v1Speed && mcdu.vRSpeed && mcdu.v2Speed && mcdu.v1Speed <= mcdu.vRSpeed && mcdu.vRSpeed <= mcdu.v2Speed;
+        return (!!mcdu.v1Speed && !!mcdu.vRSpeed ? mcdu.v1Speed <= mcdu.vRSpeed : true)
+            && (!!mcdu.vRSpeed && !!mcdu.v2Speed ? mcdu.vRSpeed <= mcdu.v2Speed : true)
+            && (!!mcdu.v1Speed && !!mcdu.v2Speed ? mcdu.v1Speed <= mcdu.v2Speed : true);
+    }
+
+    vSpeedDisagreeCheck(
+        v1 = SimVar.GetSimVarValue("L:AIRLINER_V1_SPEED", "Knots"),
+        vr = SimVar.GetSimVarValue("L:AIRLINER_VR_SPEED", "Knots"),
+        v2 = SimVar.GetSimVarValue("L:AIRLINER_V2_SPEED", "Knots")
+    ) {
+        if ((!!v1 && !!vr ? v1 > vr : false) || (!!vr && !!v2 ? vr > v2 : false) || (!!v1 && !!v2 ? v1 > v2 : false)) {
+            this.addNewMessage(NXSystemMessages.vToDisagree, this.checkVSpeedDisagree);
+        }
     }
 
     trySetV1Speed(s) {
@@ -1289,10 +1301,9 @@ class FMCMainDisplay extends BaseAirliners {
         if (isFinite(v)) {
             if (v >= 90 && v < 1000) {
                 this.v1Speed = v;
-                SimVar.SetSimVarValue("L:AIRLINER_V1_SPEED", "Knots", this.v1Speed);
-                if ((v > SimVar.GetSimVarValue("L:AIRLINER_VR_SPEED", "Knots") || v > SimVar.GetSimVarValue("L:AIRLINER_V2_SPEED", "Knots")) && SimVar.GetSimVarValue("L:AIRLINER_VR_SPEED", "Knots") !== -1 && SimVar.GetSimVarValue("L:AIRLINER_V2_SPEED", "Knots") !== -1) {
-                    this.addNewMessage(NXSystemMessages.vToDisagree, this.checkVSpeedDisagree);
-                }
+                SimVar.SetSimVarValue("L:AIRLINER_V1_SPEED", "Knots", this.v1Speed).then(() => {
+                    this.vSpeedDisagreeCheck();
+                });
                 return true;
             }
             this.addNewMessage(NXSystemMessages.entryOutOfRange);
@@ -1311,10 +1322,9 @@ class FMCMainDisplay extends BaseAirliners {
         if (isFinite(v)) {
             if (v >= 90 && v < 1000) {
                 this.vRSpeed = v;
-                SimVar.SetSimVarValue("L:AIRLINER_VR_SPEED", "Knots", this.vRSpeed);
-                if ((v < SimVar.GetSimVarValue("L:AIRLINER_V1_SPEED", "Knots") || v > SimVar.GetSimVarValue("L:AIRLINER_V2_SPEED", "Knots")) && SimVar.GetSimVarValue("L:AIRLINER_V1_SPEED", "Knots") !== -1 && SimVar.GetSimVarValue("L:AIRLINER_V2_SPEED", "Knots") !== -1) {
-                    this.addNewMessage(NXSystemMessages.vToDisagree, this.checkVSpeedDisagree);
-                }
+                SimVar.SetSimVarValue("L:AIRLINER_VR_SPEED", "Knots", this.vRSpeed).then(() => {
+                    this.vSpeedDisagreeCheck();
+                });
                 return true;
             }
             this.addNewMessage(NXSystemMessages.entryOutOfRange);
@@ -1333,10 +1343,9 @@ class FMCMainDisplay extends BaseAirliners {
         if (isFinite(v)) {
             if (v >= 90 && v < 1000) {
                 this.v2Speed = v;
-                SimVar.SetSimVarValue("L:AIRLINER_V2_SPEED", "Knots", this.v2Speed);
-                if ((v < SimVar.GetSimVarValue("L:AIRLINER_V1_SPEED", "Knots") || v < SimVar.GetSimVarValue("L:AIRLINER_VR_SPEED", "Knots")) && SimVar.GetSimVarValue("L:AIRLINER_V1_SPEED", "Knots") !== -1 && SimVar.GetSimVarValue("L:AIRLINER_VR_SPEED", "Knots") !== -1) {
-                    this.addNewMessage(NXSystemMessages.vToDisagree, this.checkVSpeedDisagree);
-                }
+                SimVar.SetSimVarValue("L:AIRLINER_V2_SPEED", "Knots", this.v2Speed).then(() => {
+                    this.vSpeedDisagreeCheck();
+                });
                 return true;
             }
             this.addNewMessage(NXSystemMessages.entryOutOfRange);